Implemented better type-lookups.
The type lookups are now completely handled by the Scope class, which
could then be extended in the future by new languages that still want
to utilize calls into the compiler.

Aside from that, it is also easier to understand what is going on, as
all lookup is done at one place, instead of partly during initialization
of the type tree.