Renamed FastRelease to FastDebug. Fixed issues with the configuration.
Previously FastRelease and Release conflicted due to them sharing the same entry point in DLLs. This caused the same DLL to be loaded twice, causing massive name conflicts.
FastDebug is now a proper debug configuration, it enables the DEBUG flag in Storm, but is built with higher optimization levels and less compiler based checks, so that Storm runs at a reasonable speed. Asserts are also enabled in FastDebug.