Cleanup of Basic Storm syntax.
Previously, packages names like "a:b[]" and "foo?" were accepted. This was due to an unintended
reuse of the Type rule for the Package rule. It may be useful to express Type as Package + Name in
the future.
Removed SStr() in Name so that the already created variable could be reused instead of requiring a
copy each time this very common rule was used. This potentially speeds up the parsing.