Possible to return values.
It is now possible to return values in Basic Storm. Do do this, the
CodeGen class has been extended to keep track of return type, and the
possibility to move parameters around inside a Frame. This also
simplifies a lot of code generating parameters in various places (not
done yet).