Simplified the ordering logic in the parser.
As the old ordering logic (execOrder) started introducing cycles in the
states of the parser, and the logic for detecting and correcting these
cycles became more and more complex, I simplified the logic a lot so
that it is easier to understand and does not introduce cycles
anymore. At least not in the cases where the old one introduced cycles,
it is not proven entirely correct.