Basic GC operation is ok.
Realized we're using 'remote references', in the GcType. These are not
supported by MPS, which means we can not fix any references that live
inside GcType. Another solution needs to be found. In this commit,
everyting works since the Type makes sure to update the
GcType (alone). This does not, however, keep the type alive properly if
not for the global type table.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>