Split Compiler/ into Compiler/ and Core/
Just like before, the runtime parts are now in Core/ while the
compiler-specific parts are located in Compiler/. This is to allow using
the core types in shared libraries further along the road. It is good to
do this split early, since it forces one to think about the interface
needed from the beginning, rather than slapping it on later on.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>