Tests for 64-bit transform.
Found and fixed a few bugs related to stack layout, register
preservation and similar. Also added templates to emacs.el, so one can
not specify a cpp.template or h.template-file which emacs will load when
a new file is created. This overrides the default generated one if a
template file exists.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>