Added proxies to instructions, fixed 64-bit cmp.
Now it is possible to write:
*to << mov(eax, edx);

instead of
*to << mov(e, eax, edx);

Which is easier to read and write.

Also fixed a hard-to-find bug in cmp for 64-bit op-codes where numbers
differing by bit # 32 failed to compare correctly due to splitting the
numbers into two. Now it uses sub and sbb internally, which perform
correctly.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>