Fixed a memory corruption bug.
The bug turned out to be caused by types derived from Type not being
properly marked as such in their GcType. This caused the types to be
allocated on the moving pool in the Gc and their GcType pointer not
being properly scanned.

In the process of finding this bug, another bug related to memory layout
was discovered. It turns out that 64-bit integers are aligned to 8 bytes
even on a 32-bit machine. Storm did not account for this previously,
causing an invalid object layout to be generated. There is now a sanity
check implemented to catch these kind of things. This is in need of some
cleanup.

Another by-product of this bug is the implementation of a malloc-based
Gc-backend which only allocates memory from malloc without ever freeing
it. This is useful for running memory check tools like Valgrind and
DrMemory, as they probably do not cooperate very well with a moving Gc
like MPS. This can be enabled by defining STORM_GC_MALLOC in
Compiler/Storm.h. DrMemory be run using 'mem_debug.sh'.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>