GC now cleans up GcType objs as needed.
Now, when a Type asks the Gc to free a GcType, it does not do so
immediatly, as objects of that type may still need to be
scanned. Instead, it puts them in a set and prunes that by
mark-and-sweep if it ever gets too large (currently 100 elements). This
requires a complete scan of the heap, which is expensive. But as we only
do it every 100 or so elements, and since removal of GcType objects are
rare, this should not be an issue.

Also fixed thread-safety in the Type object for vital functions
there. Made gcType() lazy-loaded for Types in the process.

Implemented proper matching during find() operations on Scope.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>