Added 'not' op-code. Fixed vtable-related crash.
The crash occured since we stored a non-base pointer inside code
segments when using references to vtables. This has now been solved by
hiding this fact inside VTable and let VTable generate the code that
sets the vtable of objects. This also allow us to notice when a vtable
has actually been used, which could not be done before.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>