Improved memory checking in MPS.
Rewrote the old memory debug mode to include more redundancy and a
larger header as well as a footer. May contain bugs if stressed, as I
could not find the solution to another seemingly gc-related bug using
this mode (however, the bug dissapeared in the regular mode, so I have
no clue at the moment).

Introduced 'dbg_assert' which causes a breakpoint to be hit on failure.

Various small but possibly fatal small bugfixes in the compiler.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>