Fixed a possible memory corruption in hash tables.
When a hash table needed to shrink, it reallocated the storage but
failed to reset 'lastFree'. If 'lastFree' happened to be larger than the
new smaller size we would access unallocated memory the first time
'freeSlot' was called and sometimes write there. Any subsequent calls to
'freeSlot' would most likely not cause an out of bounds access as
'lastFree' would then be properly wrapped.

This also seems to fix the issue with the CHECK_MEMORY mode in the MPS
implementation. Hopefully, there are few Gc-related bugs remaining now.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>