Array<Value> is now properly working.
Previously Array<Value> was not initialized properly as the loadAll()
function was called whenever an instance of Array<Value> was needed, and
loadAll() required Array<Value>:s to succeed. Interestingly, this did
not crash the compiler.

Now, the Array<> class tells Type that it will not add any variables, so
the loadAll() can be delayed until any member is actually used.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>