Fixed a bug in stormAs().
Basic Storm expected a boolean from stormAs, which now returns a
pointer like as<T>. This caused if (x as Foo) to falsely conclude that
'x' is not a 'Foo' sometimes due to the pointer value of 'x'.

Also improved the x # y # z - syntax so that a greater effort is made to
find an <<-operator and thus not needing a toS() function to be
generated in many more cases than before.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>