Fixed the issue causing compilation to fail.
This was since 'next' is caller earlier than anticipated for Basic
Storm, so the productions required for parsing includes were not yet
loaded, causing everything to fail. This is now remedied by letting the
readers know what Storm is going to read and thereby possibly delaying
creation of additional readers until later.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>