Added some better heuristics to File.
Now: if one adds a use-statement to the end of the use list in Storm, it is not
updated anymore. However, most of the rest works better than before.

When an incomplete start of a slide is encountered in a presentation, tab
indentation crashes the plugin on the new line for some reason.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>