Improved the re-parse logic. Fixed threading issue.
Previously, updating entire parts in the background could be aborted mid-way by
an edit operation. This caused the paused updating to proceed and replace the
contents of the part with stale data, causing us to lose characters. This is now
remedied by using a lock in the Server class, effectively preventing more than
one thing from going on at the same time. This is the model that was intended
from the beginning.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>