Rewrote the GC integration to use one object format for everything.
This was made since I had suspicions that the GC messed up when using multiple
object formats, as well as a means of introducing proper barriers to the code
allocations.

The rewrite cleaned up the pool debugging and revealed that the wrong size was
passed to mps_commit() during allocations, which probably confused MPS when
allocating strings of an odd length etc. (where the allocation size was not
aligned to start with, which is rare).

Also, found out that calling 'collect()' repeatedly on a nearly empty GC arena
seemingly causes it to make really strange decisions when scheduling garbage
collections. It does not seem to schedule collections until way later, when the
collections are expensive as the heap consists of many small segments. This is
just an hypothesis, and not yet verified. From this insight, the GC-only tests
do not cause the GC to consume a lot of time during shutdown.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>