Attempted to place buffers in a non-protected pool as per MPS docs.
Introduced a preprocessor flag in 'Gc.h' which optionally enables an additional
pool used for buffers passed to the operating system and other external
libraries. The MPS manual recommends the LO (Leaf Object) pool for this, as any
thread in the system (even threads not registered with the MPS) can access
objects in that pool (objects are not moved nor protected).

However, when using the LO pool, the MPS hits an assertion. The documentation is
also vague regarding to the properties of the LO pool. The pool summary table
states that only client pointers to objects in LO pools counts as references to
an object (ie. ambiguous references pointing inside the object might not count),
which is unacceptable as most of the Storm code assumes this to be
true. Furthermore, the pool summary table states that LO pools protect their
content, while the page detailing the pool's properties explicitly state the
opposite.

Because of this, the preprocessor flag can also specify that the pool AMCZ
should be used for buffer allocations. AMCZ allows objects to move, but as long
as they are pinned by an ambiguous reference this should not be an
issue. Furthermore, AMCZ does not protect its contents, which is good when
interfacing with the operating system asynchronously. AMCZ also seems to be
recommended by the manual as it is one of the most performant pools in the MPS.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>