Fixed crash related to the new code format.
Previous rewrites forgot to alter the CodeUpdater class to properly support the
new storage model for pointers inside of code segments. The implementation only
update the pointers actually read by the CPU, not the pointers scanned by the
GC (which are eventually used to update the aformentioned CPU-read
pointers). Most notably, this caused lazy compilation (through the LazyCode
class) to be invoked every time a function was called after a garbage collection
was triggered. This, in turn, caused almost every call to functions to dispatch
to other threads, causing very many stacks to be allocated until no more memory
was available. An appropriate exception is thrown when spawning a thread without
memory to back it up.

Aside from this, the entry-point name for shared libraries are now correct even
in the 'slow debug' configuration, so that Storm does not attempt to load two
versions of the same shared libraries.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>