Found the performance hog in the new error recovery.
It is caused by excessive calls to 'limitedReduce' and this can be helped by not
interleaving calls to 'reduceAll' with calls to 'shiftAll', and instead only
call (a stronger version of) 'shiftAll' at the end of 'advanceAll'. This
requires 'shiftAll' to be able to skip over nonterminals and in a way perform
simple reductions (productions with zero entries).

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>