Fixed issues in the re-parsing logic.
Parsing of strings (among other things) are still slow since the language server
thinks it is getting bad results from the parser and tries to parse many times
on a single operation. The re-parse logic should probably be re-written to
remedy this so that the language server first traverses the tree to decide where
it should re-parse and then do that once. If that fails (or the result is too
bad for some reason), a further re-parse of more or less the entire file should
be scheduled for later.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>