Changed penalization system for errors in the GLR parser.
Now, 'skipped' is renamed to 'chars' in InfoErrors to indicate that 'chars'
represent the number of characters inside the erroneous productions. As such
'chars' give an approximate measure of how many characters are affected by an
error recovery. This seems to work better than previously with regards to long
comment sections, as these comment sections typically only have two or three
shifts while they affect many characters. The desired correction typically shows
about 10 shifts but only a handful (compared to the comment solution) characters
affected.

This also made the previous solution where regexes in multiline comments did not
match newline characters work worse than simply matching as much as possible in
regexes. This made it easier for the error recovery to backtrack to the correct
place, and thereby produce a good result for the parsing. As such, multiline
comments in Java and Storm matches newlines where possible.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>