Updated the re-parse logic in the language server.
Now, the language server is more aggressive when looking at what to re-parse in
order to only call the parser once. The GLR parser is generally very fast as long as
it does not have to do error recovery, and thus calling it once with a bit more
context will yield a good result faster than calling it a few times around the
erroneous area (forcing it to do error recovery multiple times) in hope to save
time by parsing fewer characters.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>