Improved re-parsing logic to include more environment around errors.
Without taking boundaries into account, we could get stuck into local
optimum. For example, it might have been beneficial to pretent something was a
part of an expression at one point in the input, which caused a big comment
block to be included as expressions as well. On further edits, only the
'expression' was re-parsed as it was long enough, causing the previous error to
remain. Now, we require that the re-parsed substring extends beyond the
erroneous region by a few characters, which hopefully makes the parser do better
decisions in future re-parses at least.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>