mpsMakeFwd does not alter the object forwarded to anymore.
Code segments need to have their pointers updated whenever they are
moved. This was previously done in the mpsMakeFwd function by altering the
object pointed to by the 'to' parameter. This worked, but is possibly breaking
the interface of the function. Now, we make sure that moved objects are scanned
once more by including a pointer to the code segment itself in the code
segment. This requires MPS to scan the object again as the self pointer needs to
be updated by MPS_FIX.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>