Fixed broken typechecking when initializing variables.
Previously, it was possible to do the following:

Float x = Size(2, 3);

which caused memory to be overwritten. This was possible with all built-in types
on the left hand side. This was possible due to a missing type-check in the case
of built-in types. Now, the compiler properly rejects this kind of code.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>