Fixed a bug when using 'fnParamRef'.
Previously, the code generation backend used the size of the source reference
instead of the size of the actual reference when computing stack offsets of
function calls when 'fnParamRef' was used. This caused issues when multiple
reference params were used, which in turn caused crashes. This is now fixed, and
a test case illustrating the bug was created.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>