Researched member pointers on GCC-X64.
It turns out that GCC avoids a vtable stub by storing tagged vtable offsets, at
least on X86-64. Updated the FnCall interface to properly handle this.

This also showed the reason for an earlier bug. The generated code for indirect
member calls wrongly identified the doEndDetour function as a member function
since it was not properly aligned by default. This is now fixed, and the UThread
spawn mechanism will perform a vtable lookup if required. This is not yet tested though.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>