Semaphore class now works properly even when signals arrive.
Previously, a thread waiting in Semaphore::down() would wake early when a signal
was dispatched to that thread. This would cause all kinds of strange threading
problems, especially since the MPS uses signals to stop and resume threads. Now,
this case is properly taken care of, which allows all stress test to
pass (except GcThreadTest3, which requires code generation).

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>