Parameter layout for X86-64 seems to work.
The implementations matches the observations made when studying GCC in
action. However, this does not match my interpretation of the System V ABI
entirely. This is most likely due to the System V ABI being slightly vague on
some corner cases. Anyway, if GCC is "wrong", we want to be "wrong" too in order
to be compatible with GCC!

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>