Revised the previous changes to code::Size.
The previous changes broke the ability to properly extract the
current (unaligned) value from a code::Size object, which is required to perform
proper layout. Now, operator+(Offset, Size) has been updated to match the
behavior of first casting the Size object to an Offset object and then performing
the addition. This might break things, however, and as such they have been
marked with TODO().

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>