Returning floating point values working properly for X86-64.
This required implementing basic support for the xmm registers. In the future,
it might be viable to use XMM for floating point operations on X86 as well. This
would lead to a nicer instruction set for floating point values in the
IR. However, being able to mix floating point and integer registers freely could
cause problems, as fp-registers and integer registers are mostly disjoint.

Also fixed returning complex structs.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>