Storm generates TypeDesc objects for Types. Pure functions supported.
Implemented support for generating meaningful TypeDesc objects. In the process,
it was also neccessary to know if constructors and destructors are
trivial. Therefore, the concept of 'pure' functions was introduced. A pure
function is not expected to have side effects (much like default constructors
and destructors in C++). This can be used to perform constant folding in a later
stage of the compiler. Then we can fold all functions that are pure, not just a
special set of built in functions. Eventually, we want to support marking
functions as pure in Storm and in C++.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>