Fixed issues causing the BSException test to fail.
This was caused by an error in the layout of spilled register parameters in the
X64 backend. This caused two integer parameters to be written to the same
location, thereby trashing one of the values.

Also, no CFA information is emitted for epilogs anymore, as that would require
a dynamic amount of memory to be allocated for the CFA records. This should not
be a problem as long as no exceptions are thrown on the two last instructions of
the epilog or so.

The epilog CFA information also requires additional care if the epilog is not
the last thing in the function. If the epilog was used for a return statement,
the remaining code would not have proper CFA information, causing the unwinding
to fail.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>