Fixed a crash when calling the ctor-wrappers on threads.
Since the createN wrapper functions were not member functions, they were not
always aligned properly. When calling these on a thread, the spawn functionality
was informed that we're dealing with a member function and therefore any odd
pointers were treated as vtable offsets (which is not correct in this
case). Now, all the createN templates have proper alignment, so the code works
as intended.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>