Improved the UThread detour logic.
Previously there was a limit of one active detour for each active thread in the
system. Since calling copy constructors could trigger complex behaviours, such
as thread switches or other syncronization, this may not always be enough in the
general case.

Therefore, the global limit is lifted by putting the detour information inside
the UThreadData (a single UThread) instead of the UThreadState (one for each OS
thread), so that each thread can remember which thread they are a detour
from (if any). The effect of this is that each thread can be replaced (possibly
multiple times) by another detour thread in the scheduler. Thereby, code
executing as a detour is not treated special in that it is unable to initiate
yet another detour. Furthermore, we properly handle the case where a detour
calls UThread::leave() (perhaps by using a synchronization primitive), and the
newly scheduled thread tries to initiate yet another detour (probably by calling
UThread::spawn()), which is likely to happen in practice in larger systems.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>