Made function alignment global on Linux.
It seems like __attribute__((aligned(2))) needs to be placed either first or
last in function declarations to make it affect the function and not the return
value on GCC. This means that it is not possible to use the CODECALL macro in
its current form to mark functions that require more alignment on
Linux. Therefore, the flag -falign-functions=2 was added to enforce a slightly
higer function alignment globally.

This makes the parsing demo work properly on Linux.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>