Using SA_RESTART in the signal handling inside the MPS.
If it is not used, some libraries (such as Gtk+) may experience a EINTR result from system calls
such as read() and write(), which they do not handle properly. Gtk+, for example, will print a
message stating that it received a EINTR and it does not know what to do with it. As far as I can
tell, there is no good way of recovering from this error, so restarting system calls should be a
viable fix for that.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>