Removed some flickering when resizing Gtk+ windows with painters.
There is still the issue where animations freeze. The most reliable way of triggering the issue
seems to be to resize the window a lot. It looks like the issue boils down to the IOCondition not
being properly woken by a call to signal(). When debugging, GDB usually shows the render thread
blocked inside poll() while the 'signaled' flag is set to 1. Removing the use of 'signaled' seems to
relieve the issue to some degree at least.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>