Fixed a threading issue in the IOCondition under Linux.
Previously, there was a race condition after clearing the 'signaled' flag and
reading from the eventfd. If another thread happened to call 'signal' right
after the 'atomicWrite' but before the call to 'read', a thread would generate a
new signal on the eventfd that would be consumed along the old signals during
the call to 'read'. Since 'signal' sets the flag once more, this would cause any
additional calls to 'signal' to do nothing, even though nothing was signaled on
the eventfd, which causes a deadlock.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>