Fixed a use before initialization bug in the threading system.
In rare cases, the GC could see UTreads in the process of initialization (UThreadStack
objects). This made the GC consider them as being the main thread, which would subsequently lead to
a crash. This is fixed by adding a flag to the UThreadStack objects, so that the GC detects this
case and just ignores those threads until they are properly initialized.

Furthermore, UThreads that were used as the target of a detour operation was previously scanned by
the wrong OS thread. This could also in rare cases cause the GC scanning to think two UThreads were
running at the same time. This is fixed now by making the detour target being 'adopted' by the
proper OS thread during the detour. This is done with two additional variables inside UThreadStack.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>