Tried disabling signals while calling into Gtk+:s main loop.
It seems that the EINTR during startup usually comes from the call to g_main_context_prepare. So,
tried to block all signals during that call (even the pause call from the GC, which could be
dangerous). It seems to vastly improve the situation with Gtk+. It could introduce deadlocks wrt the
GC though.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>