Cairo is compiled and used as a shared library.
Cairo is compiled as a separate shared library intended to be distributed
alongside the Gui so-file on Linux. To ensure that our version of Cairo is
loaded, the RPATH of the Gui library is set to $ORIGIN.

Since the Cairo so-file has to be named libcairo.so.2, Storm now discards any
numbers at the end of file extensions. Therefore, libcairo.so.2 is loaded as an
so-file would be instead of generating a warning.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>