Resizing the existing cairo surface instead of re-creating it.
Cairo provides a function to resize GL surfaces, which was not previously
used. Now, we resize the surface instead of destroying it and re-creating it
with a new size. This lets us re-use the cairo context as well, which saves
quite a lot of allocations/deallocations when resizing a window.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>