Fixed widget re-creation on older Gtk+.
Previously, re-creating (or rather re-realizing) a widget when running Gtk+3.10
caused it to be drawn properly, but it was impossible to interact with such
widgets. This is solved now, even though older Gtk+ will emit a warning during
the re-creation process. This is not the case with Gtk+3.22, however, so it is
an acceptable solution since changing painters is probably not very commonly
used.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>