Fixed a bug that caused more or less random crashes in Linux.
This was caused by an error in code::x64::singleInt(). It erroneously thought that values between
0x80000000 and 0x100000000 could fit inside a 32-bit signed integer. This caused problems for
example when two functions were approximately 2GB apart from each other in memory (which could
happen at any time due to GC activity), when the GC implementation would use a near jump instead
of a far jump. This, in turn, results in an erroneous jump target and a crash.

Also modified GcCodeRef::relative so that it does not scan pointers, as the comment suggests.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>