Fixed issues when interacting with the Player class.
If Player.stop() was called while the player was filling a buffer, two UThreads
could accidentally try to read from the sound source at the same time. This
confused the underlying library quite a lot, which could cause crashes (eg. for
FLAC and MP3 files) in the worst case. This is fixed by adding a lock to the
Player class, locking all accesses to the sound stream.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>