Fixed crash while running BF code on Linux.
This was due to the BF backend used push() and pop() instructions for temporary
storage, which caused the stack to be mis-aligned when calling into the string
buffer. This sometimes caused a segmentation fault, eg. when the GC tried to
make new allocations and called memcpy() on large enough memory.

The BF backend was modified to save its state to a proper variable, which solved
the problem. In the process, some of the now unneeded Operand() casts were
removed.

While investigating, some missing register names in the X64 backend were
added. Namely, the registers r8l-r15l, which were previously missing. These are
sometimes used by backends, so it is annoying when they can not be shown in the
textual output. The second parameter to arena.transform() function was also
declared as MAYBE() to allow using it to display output from
transformations. Currently, the virtual dispatch for that function is broken
for some reason. Further investigations are necessary.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>