Function names are shown for all Storm functions on Windows.
Now, Windows uses a function table similar to the DwarfTable used on X86-64 to
look up the start of the current function while an exception is thrown. This
lets us look up the Binary object for all functions, not only the functions that
were using exception handling.

This means that the Binary that is being pushed on the stack is not needed anymore.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>