Moved the CodeTable into Code/.
It can now be used as a global variable, much like the DwarfTable for X64. This
lets us use it for exception handling as well.

Also replaced the call to std::sort with a heapsort (std::make_heap,
std::sort_heap) to make sure that std::sort does not crash if any pointer would
change while we are sorting the array.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>