Reduced memory usage for functions on X86.
Previously, the owning Binary object was stored twice due to the new stack trace
ability. This is now fixed by storing a pointer to the start of the function on
the stack instead of to the Binary object. The Seh can then use that information
to locate the Binary object through the same pointer used by the unwinding.

Also, X86 metadata is properly aligned so that stack unwinding should be
slightly faster.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>