Fixed a bug causing VTable lookup for C++ objects to fail.
When deciding to use VTable lookup for a function, the Type class first adds the
base class and then the derived class. This means that the function in the base
class first thinks that VTable lookup is not required. When the derived function
is added, the code failed to forcefully set the lookup of parent functions when
they were using a C++ VTable since a different code path was used.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>