Introduced a 'scope' parameter to 'find'.
This makes sure the entire system are respecting the visibility of named
objects. Some parts of the system are bypassing these checks by passing Scope(),
but this is only done where it is reasonable, eg. during package loading, and
when computing vtables.

There is some issues with the automatic toS function and vtables that are
causing tests to fail. This should be investigated, and the feature should be
properly tested and implemented in Basic Storm.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>