Basic Storm 'generates' comparison operators based on < and ==.
Basic Storm automatically falls back on more 'basic' comparison operators if the
desired one does not exist. This gives the illusion that the system generates
the required operators on demand (which we might want to do in the future). This
means that it is only necessary to implement <. It is also nice to implement ==,
as the == operator is usually faster to compute than calling < twice. Since this
is a feature of Basic Storm, a Handle will not automatically get a == operator
if < is defined, but the helper function 'equal' uses '<' if '==' is not
present. This does not work if the user defines '>' even though it works
properly in Basic Storm.

Furthermore, the issue where expressions like '5 / 2.0' did not compile properly
in Basic Storm.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>