Implemented a working wrapper around STL iterators for Handle types.
However, this implementation currently assumes that all object instances are
stored on the stack, which may not always be the case for eg. std::sort.

Maybe it is easier if we implement our own variant of std::sort. Then we do not
need to worry about different implementations having different usage patterns
that we need to account for.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>