Fixed a bug with function calls on X86.
This bug caused the backend to not preserve registers when using
eg. 'ptrRelative' for a reference parameter to a function call.

Rewrote the 'callRaw' mechanism for function pointers to utilize the new
possibilities in the backend. This likely breaks the Linux port until the same
issue is fixed in the X64 backend.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>