Improved generation of documentation.
Documentation is now generated for the entire compiler by default. The
documentation is placed in a file named 'doc' next to the shared library
itself (or directly in root/ in the case of the compiler). The generated file is
deterministic, so using the same file for multiple builds should not be an
issue.

Storm was also improved to properly handle (ie. ignore) empty files. Previously,
it was confused as the paths 'foo/bar' and 'foo//bar' are usually considered
equal by the OS. The Url class will now signal this error, and package loading
will refuse to generate names for files without an extension.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>