Created Doc classes for holding documentation in Storm.
The information is filled in from the information in CppTypes. However, no
support for loading the generated documentation files is implemented yet.

The name convention for documentation files have changed from 'doc' into
'LibName_doc' to make it possible to have more than one dynamic library in a
single package. The filename containing the documentation is stored in the
CppWorld structure (relative to the library).

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>