Fixed crash when evaluating 10 + foo(), where foo returns void.
This was caused by unclear nomenclature in the ExprResult class. The members
'any' and 'empty' are now replaced with 'value', 'empty' and 'nothing' which
more clearly indicate what kind of value is contained in the result. Now, it is
not as easy to mistakenly think that '.any()' implies that the returned result
is not 'void' and crash.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>