Implemented support for 'setters' or 'assignment functions'.
If a function is declared as 'assign' in Basic Storm or 'STORM_ASSIGN' in C++,
Basic Storm allows using those functions as an 'assignment target', ie. it is
possible to use 'assign foo(Int x)' as 'foo = 10'. This makes it possible to
implement getters and setters that act exactly like regular variables. At least
in Basic Storm.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>