Converted from using a std::set to a proper storm::Set Debug.cpp.
Previously, we used a std::set, which was unaware of moving objects inside the
object traversal code in Compiler/Debug.cpp. This could in theory fail at
arbitrary points, and cause other GC havoc. This is now remedied by using a
proper set from Storm that properly respects the oddities of having a moving GC.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>