Updated the Basic Storm documentation wrt actors.
The documentation on types is now more clear about the pitfalls when working with actors: mainly
that things are copied automatically at times where it might not be obvious. Furthermore, it is not
possible to alter data members in actors belonging to other threads directly. It is necessary to use
functions for that.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>