Modified the Gtk interaction to reduce stuttering.
When using the mode GTK_MT_HW_COPY (ie. copying on the Ui thread without coordinating with Gtk+),
some systems experienced fairly long (about 500 ms) freezes in animation. These stutters were not
caused by blocking calls in the code (the underlying Ui and Render threads were running normally
during stutters), but more likely some kind of race condition with Gtk+. Therefore, the mode was
changed to GTK_MT_HW_CAIRO to copy to the Gtk+ surface using Cairo, which solved the problem. The
major downside with this approach is higher CPU utilization and slightly more VSync artifacts.

Note: The issue could still be in the handling of double surfaces, but that requires further
investigation. It would be nice to be able to use GL while bypassing Gtk+ (since it usually does
software rendering).

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>