Allow simple 'traversals' in the syntax language.
A traversal is when one accesses member variables or calls simple functions in
an actual parameter to a function call or to a rule invocation in the syntax
language. This can be used to simplify productions in some cases, such as when
implementing nested types in Basic Storm. This previously required a dummy
production matching the empty string to access a member of bs::Class.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>