It is now possible to document bnf files as well.
The backend for reading syntax definitions now properly understands which
comments could be documentation and appends them to the name tree just like
Basic Storm does.

Also fixed a bug in the parsing of comments (mainly for C++) that would consider
two distinct blocks of single-line comments as if it was one block. Thus the
following two comments would both be considered to be documentation for foo()
below.

// First comment

// Second comment
int foo();

This is no longer the case, and parsing behaves as one would expect.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>