Resolved an issue that caused out-of-order loading sometimes.
The issue was that invisible dependencies were created between all types sharing
a base class (ie. most types in the system) due to eager creation of VTables for
types. This could cause a seemingly unrelated type being forcibly loaded just so
the system could figure out if vtable dispatch should be used for a function in
an unrelated type. This issue is resolved by making sure to use the new
'tryFindHere' function when traversing child classes to examine VTables. The
system ensures proper vtable creation by calling 'forceLoad' when the VTable is
retrieved from a type instead.

This problem appeared as the system failing to find some types that should
exist. These types were commonly defined in different files (eg. syntax + basic
storm) and in the same package. The issue usually appeared since a syntax node
was loaded before the types defined in basic storm were loaded. This appeared
randomly since the traversal order of children is dependent on addresses, which
vary.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>